[scanner] refactor: split NamespaceManager into focused sub-components and hooks
Extracts NamespaceAccessPanel, NamespaceFilterBar, NamespaceClusterGroup sub-components
and useNamespaceFetch custom hook from NamespaceManager.tsx into focused modules.

- NamespaceManager.tsx reduced from 925 to 353 lines (-62%)
- useNamespaceFetch.ts: custom hook encapsulating namespace fetching logic,
  caching, progressive loading, and auto-refresh (~320 lines)
- NamespaceFilterBar.tsx: presentational component for search and group-by
  controls (~60 lines, zero hooks)
- NamespaceClusterGroup.tsx: presentational component for per-cluster
  collapsible groups (~120 lines)
- NamespaceAccessPanel.tsx: component managing access entries and revocation
  (~150 lines, 3 hooks)

Each extracted component has a clear, focused purpose and is independently testable.
Parent component now acts primarily as a composition layer.

Fixes #20601
Fixes #20611
Fixes #20612
Fixes #20613
Fixes #20615
